+++ back-poly/polyrate/src/main/java/nick/project/polyrate/Transacao/Repository/TransacaoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.time.LocalDateTime;
 
-public interface TransacaoRepository extends JpaRepository<TransacaoEntity, BigInteger> {
+public interface TransacaoRepository extends JpaRepository<TransacaoEntity, Integer> {
 
    Page<TransacaoEntity> findByCreatedAtBetween(LocalDateTime inicio, LocalDateTime fim, Pageable pageable);
 }
